Moved tile actors to seperate file, Added back player to level

+++ Black Box Engine/Graphics/Camera.h
@@ -9,8 +9,8 @@ namespace BlackBoxEngine
     class Camera
     {
     public:
-        int m_screenXPos = 0;
-        int m_screenYPos = 0;
+        float m_screenXPos = 0;
+        float m_screenYPos = 0;
         float m_width   = 256;
         float m_height  = 256;
